+++ BlankCoreAppCopyTask/Services/SynchronizationOneThread.cs
@@ -20,7 +20,8 @@ namespace BlankCoreAppCopyTask.Services
             _hashCalculator = hashCalculator;
         }
 
-        public async Task Copy(ImmutableArray<IFileToCopy> files, Action<double> updater)
+        public async Task<Result<IFileToCopy>[]> Copy(ImmutableArray<IFileToCopy> files, Action<double> updater,
+            CancellationToken sourceToken)
         {
             var sum = GetSumOfAllFileSize(files);
 
@@ -30,12 +31,18 @@ namespace BlankCoreAppCopyTask.Services
                 Interlocked.Add(ref copyProgressInfo, bytesTransferred);
                 Dispatcher.CurrentDispatcher.Invoke(() => { updater?.Invoke((double) copyProgressInfo / sum); });
             });
+            var bag = new ConcurrentBag<Result<IFileToCopy>>();
             var copyOperation = Task.Run(async () =>
             {
-                foreach (var file in files) await Copy(file, action);
-            });
+                foreach (var file in files)
+                {
+                    var result = await Copy(file, action, sourceToken);
+                    bag.Add(result);
+                }
+            }, sourceToken);
 
             await copyOperation;
+            return bag.ToArray();
         }
 
         public async Task<ImmutableArray<IFileToCopy>> CreateListOfFilesToCopy(string src, string dst)
@@ -70,13 +77,36 @@ namespace BlankCoreAppCopyTask.Services
             return files.Sum(file => file.Size);
         }
 
-        private async Task Copy(IFileToCopy file, Action<long> progressUpdate)
+        private static async Task<Result<IFileToCopy>> Copy(IFileToCopy file, Action<long> progressUpdate,
+            CancellationToken sourceToken = default)
         {
-            await using var sourceStream = File.Open(file.Path, FileMode.Open);
-            if (!File.Exists(file.Destination))
+            try
             {
-                await using var destinationStream = File.Create(file.Destination);
-                await sourceStream.CopyToAsync(destinationStream, 16384, progressUpdate, CancellationToken.None);
+                if (sourceToken.IsCancellationRequested)
+                {
+                    sourceToken.ThrowIfCancellationRequested();
+                }
+
+                await using (var sourceStream = File.Open(file.Path, FileMode.Open, FileAccess.Read, FileShare.Read))
+                {
+                    if (!File.Exists(file.Destination) && sourceStream != null)
+                    {
+                        await using var destinationStream = File.Create(file.Destination);
+                        await sourceStream.CopyToAsync(destinationStream, ISynchronization.BUFFER_SIZE, progressUpdate, sourceToken);
+                    }
+                }
+
+                return new Result<IFileToCopy>(file, null);
+            }
+            catch (FileNotFoundException e)
+            {
+                Debug.WriteLine(e);
+                return new Result<IFileToCopy>(file, new Error($"file not found: {e.Message}"));
+            }
+            catch (IOException e)
+            {
+                Debug.WriteLine(e);
+                return new Result<IFileToCopy>(file, new Error($"io error: {e.Message}"));
             }
         }
     }
